% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{BRT}
\alias{BRT}
\title{A Boosted Regression Tree model to predict the spread}
\format{A caret prediction model}
\usage{
BRT
}
\description{
A Boosted Regression Tree model to predict the spread
}
\seealso{
\code{\link[SpatialBall]{ShotSeasonGraph}}
}
\keyword{datasets}
