% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{court}
\alias{court}
\title{A raster grob used to plot the court in shot charts}
\format{A rastergrob, grob y gDesc to use for ploting the court.}
\usage{
court
}
\description{
A rastergrob, grob y gDesc to use for ploting the court.
}
\seealso{
\code{\link[SpatialBall]{ShotSeasonGraph}}
}
\keyword{datasets}
