% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{OffShotSeasonGraphPlayerEight}
\alias{OffShotSeasonGraphPlayerEight}
\title{plot the shot chart of a player for an NBA Season with eight bins}
\usage{
OffShotSeasonGraphPlayerEight(Seasondata, player, nbins = 8, quant = 0.4,
  type = "PPS", MAX_Y = 418)
}
\arguments{
\item{Seasondata}{The information of shots, it can be downloaded with function
read_season}

\item{player}{the player you which to plot the shot charts of}

\item{nbins}{The number of bins the hexplot for the shot charts are made
(default is 8)}

\item{quant}{A number between 0 and 1, it determines quantile of shots used
to plot the shot chart, (default is 0.4)}

\item{type}{A character to specify if the shot chart is based on Points per
Shot ("PPS") or percentage ("PCT")}

\item{MAX_Y}{a numeric that limits the y axis of the shot chart, defaults at
280}
}
\value{
a ggplot object plotting the defensive shot chart of a given team on
an NBA season
}
\description{
This function takes an NBA season object and makes a shot chart of all the
shots takes through that regular season.
You can choose to either plot the results based on Points per Shot or on
Shooting Percentage
}
\examples{
data("season2017")
#Examples with several players
OffShotSeasonGraphPlayer(season2017, player = "Stephen Curry")
OffShotSeasonGraphPlayer(season2017, player = "DeAndre Jordan")
OffShotSeasonGraphPlayer(season2017, player = "DeMar DeRozan")
OffShotSeasonGraphPlayer(season2017, player = "Isaiah Thomas")

 #Examples with percentage instead of points per shot
OffShotSeasonGraphPlayerEight(season2017, player = "Stephen Curry", type = "PCT")
OffShotSeasonGraphPlayerEight(season2017, player = "DeAndre Jordan", type = "PCT")
}
\seealso{
\code{\link[SpatialBall]{OffShotSeasonGraphTeam}}

\code{\link[SpatialBall]{ShotSeasonGraph}}
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>
}
