% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PointShotSeasonGraph}
\alias{PointShotSeasonGraph}
\title{plot all the shots taken on an NBA Season separating makes and misses}
\usage{
PointShotSeasonGraph(Seasondata)
}
\arguments{
\item{Seasondata}{The information of shots, it can be downloaded with function
read_season}
}
\value{
a ggplot object plotting all the shots taken on an NBA season
}
\description{
This function takes an NBA season object and plots as circles all the shots
taken on an nba season, ploting makes and misses as different colors
}
\examples{
data("season2017")
PointShotSeasonGraph(season2017)
}
\seealso{
\code{\link[SpatialBall]{ShotSeasonGraph}}
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>
}
