% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PointShotSeasonGraphTeam}
\alias{PointShotSeasonGraphTeam}
\title{plot all the shots taken on an NBA Season separating makes and misses}
\usage{
PointShotSeasonGraphTeam(Seasondata, team)
}
\arguments{
\item{Seasondata}{The information of shots, it can be downloaded with function
read_season}

\item{team}{The team the shots should be graphed}
}
\value{
a ggplot object plotting all the shots taken by a team on an NBA
season
}
\description{
This function takes an NBA season object and plots as circles all the shots
taken on an nba season, ploting makes and misses as different colors
}
\examples{
data("season2017")
PointShotSeasonGraphTeam(season2017, "Hou")
PointShotSeasonGraphTeam(season2017, "Orl")
}
\seealso{
\code{\link[SpatialBall]{ShotSeasonGraph}}
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>
}
