% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{Seasonscrape}
\alias{Seasonscrape}
\title{Gets a season dataset from the NBA stats page}
\usage{
Seasonscrape(season = "2017-18", type = "Regular Season", Start = NULL,
  End = NULL)
}
\arguments{
\item{season}{the season of which you want to get the data, in a character format,
defaults to "2017-18"}

\item{type}{One of: "Regular Season", "Pre Season", "Playoffs", "All-Star",
"All Star", "Preseason", defaults to "Regular Season"}

\item{Start}{a character in a month/day/year format, if NULL (default)
it starts at the beginning of the season}

\item{End}{a character in a month/day/year format, if NULL (default)
it ends scraping at the end of the season}
}
\value{
A season database
}
\description{
Creates a season dataset
}
\examples{

\dontrun{

#Without date limits

Seasonscrape(season = "2017-18")

#With date limits

Seasonscrape(season = "2017-18", Start = "12/05/2017", End = "12/07/2017")

#Scraping playoff games with date limits

Seasonscrape(season = "2016-17", type = "Playoffs", Start = "03/25/2017", End = "03/30/2017")

}
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>
}
