% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{TeamStats}
\alias{TeamStats}
\title{Calculate Offensive and Defensive stats for teams}
\usage{
TeamStats(Seasondata)
}
\arguments{
\item{Seasondata}{The information of shots, it can be downloaded with function
read_season}
}
\value{
a list of two dataframes, one with the offensive team stats and the
other with defensive team stats
}
\description{
This function takes an NBA season object and calculates the Offensive and
Defensive stats for teams
}
\examples{
\dontrun{
data("season2017")
TeamStats(Seasondata = season2017)
}
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>
}
